# Preview: Optimize the advanced detector further by 50%