--- a/src/contacts/advanced/detector.cpp
+++ b/src/contacts/advanced/detector.cpp
@@ -58,11 +58,6 @@ BlobDetector::BlobDetector(index2_t size, BlobDetectorConfig config)
     , m_gf_window{11, 11}
     , m_touchpoints{}
 {
-    std::less<alg::wdt::QItem<f32>> op {};
-    std::vector<alg::wdt::QItem<f32>> buf {512};
-
-    m_wdt_queue = std::priority_queue { op, buf };
-
     alg::gfit::reserve(m_gf_params, 32, size);
 }