This is surprising. I thought I could only gain 5%. Nobody, not even me thought this would happen. Instead we gain another 1000us giving an 50% improvement.

***This PR isn't planned to be merged any time soon***, unless someone really insists. The fact that I didn't bother cleaning up the code proves this. The purpose of this PR is just to show that it's possible to get this size of improvements without changing the algorithm.

Looking at the perf graph, `MyQueue::{push,pop}` is 20%, so this seems to be the end of the road for this type of optimization. I had been doing architecture-preserving, function-only optimizations that didn't touch globally-passed-around data structures much. As I mentioned in the other PR, we might need to consider fixed-length data structures that trade flexibility for performance.

CPU dropped from 45% to 36% in iptsd, but still spikes to 100% when using 10 fingers. The next step I'm thinking of is trying to remove recursion/branching but that might greatly change the essence of the algorithms.

# Before

See #113

```
[22:20:12.187] [info] Vendor:       045E
[22:20:12.187] [info] Product:      099F
[22:20:12.187] [info] Buffer Size:  7487
[22:20:12.187] [info] Metadata:
[22:20:12.187] [info] rows=44, columns=64
[22:20:12.187] [info] width=25978, height=17319
[22:20:12.187] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[22:20:12.187] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[22:20:16.228] [info] Ran 2010 times
[22:20:16.228] [info] Total: 4031356μs
[22:20:16.228] [info] Mean: 2005.65μs
[22:20:16.228] [info] Standard Deviation: 674.57μs
[22:20:16.228] [info] Minimum: 144.730μs
[22:20:16.228] [info] Maximum: 3357.086μs
```

# After

```
[14:39:40.948] [info] Vendor:       045E
[14:39:40.948] [info] Product:      099F
[14:39:40.948] [info] Buffer Size:  7487
[14:39:40.948] [info] Metadata:
[14:39:40.948] [info] rows=44, columns=64
[14:39:40.948] [info] width=25978, height=17319
[14:39:40.948] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[14:39:40.948] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[14:39:42.951] [info] Ran 2010 times
[14:39:42.951] [info] Total: 1992742μs
[14:39:42.951] [info] Mean: 991.41μs
[14:39:42.951] [info] Standard Deviation: 324.57μs
[14:39:42.951] [info] Minimum: 86.030μs
[14:39:42.951] [info] Maximum: 1635.113μs
```

![image](https://user-images.githubusercontent.com/25646384/221275890-54d73271-2db8-46fe-a391-90d19d6ed660.png)