@qzed 2023-02-24
Me assuming that STL implementations would be at least somewhat optimized is the reason I went with those in the first place... (that and me being lazy).