@qzed 2023-02-24
> The strange thing is that inlining isn't always a good thing.

From what I know that could be a cache thing? Inlining usually makes code larger, so there's some balance between function calls and fitting the most frequent stuff into the cache. Even if the function is not inlined, it's likely in the cache so reasonably quick to call.

> The thing with STL is that they can't make as many assumptions as you can when you're the same person consuming the class. STL uses references everywhere, where we know the heap contains copyable type.

Right, that's still a big performance improvement. Way bigger than I'd have thought.

> Part of the improvement in this PR is from changing array-of-structures `alg::wdt::QItem<f32> data[QUEUE_MAX]` into struct-of-arrays `index_t ind[QUEUE_MAX]; i32 cost[QUEUE_MAX];`

Ah yes, I suppose that make sense.

> When I thought of replacing `std::vector` with arrays, it wasn't because of the `malloc` time. It was because the data should be stored in on contiguous location, not 10+ places scattered across memory.

Right, I kind of thought that this would have a pretty minimal effect since the `std::vector` data itself is contiguous. Vector vs. linked list, definitely. But I guess it might make some sense if we have a bunch of small vectors.

> 50% of the time is spent inside `evaluate`. It is heavily branching and I can't vectorize it. Now I think I've reached the limit of this type of optimization as I don't see how to improve that function.

Yeah, that's a bit of a problem with the wdt... which I think is also more or less the core problem of the whole "advanced" processor.